Move and rename the windows and linux templates
