+++ .teamcity/Ribasim_Windows/Project.kt
@@ -5,6 +5,7 @@ import Ribasim_Windows.buildTypes.Windows_TestDelwaqCoupling
 import Ribasim_Windows.buildTypes.Windows_TestRibasimBinaries
 import Templates.GithubCommitStatusIntegration
 import Templates.GithubPullRequestsIntegration
+import Templates.WindowsAgent
 import jetbrains.buildServer.configs.kotlin.BuildType
 import jetbrains.buildServer.configs.kotlin.FailureAction
 import jetbrains.buildServer.configs.kotlin.Project
@@ -19,6 +20,8 @@ object Project : Project({
     buildType(Windows_BuildRibasim)
     buildType(Windows_TestDelwaqCoupling)
     buildType(Windows_TestRibasimBinaries)
+
+    template(WindowsAgent)
 })
 
 object Main : BuildType({
