+++ .teamcity/Templates/WindowsAgent.kt
@@ -1,15 +1,11 @@
-package Ribasim.buildTypes
+package Templates
 
 import jetbrains.buildServer.configs.kotlin.*
 
-object Windows_1 : Template({
+object WindowsAgent : Template({
     id("Ribasim_Windows")
     name = "Ribasim_Windows"
     description = "Template for agent that uses Windows OS"
-
-    vcs {
-        cleanCheckout = true
-    }
 
     requirements {
         contains("teamcity.agent.jvm.os.name", "Windows", "RQ_422")
